Added rollDice() function.
Declares two of the six ImageViews for testing and the total textview.

Updates one or two of the dice images, dependant on the users choice on first page.

Can't find a neater way (right now) to run through the different options of 1-6 except using a when statement for each dice.