Added new Activity.
When the user selects the amount on spinner, it carries the value over and stores in "amount"

In the next activity, this value is available to the program allowing the program to select up to two die currently.

It sets the default value of "amount" to 1 if the user does not pick a spinner option.